refactor: Split useDrawing hook into layered architecture
