+++ client/src/hooks/canvas/useDrawingOperation.ts
@@ -1,0 +1,126 @@
+import { RefObject, useCallback } from 'react';
+import { DrawingData, Point, StrokeStyle } from '@troublepainter/core';
+import { useDrawingState } from './useDrawingState';
+import { RGBA } from '@/types/canvas.types';
+import { getCanvasContext } from '@/utils/getCanvasContext';
+import { hexToRGBA } from '@/utils/hexToRGBA';
+
+const fillTargetColor = (pos: number, fillColor: RGBA, pixelArray: Uint8ClampedArray) => {
+  pixelArray[pos] = fillColor.r;
+  pixelArray[pos + 1] = fillColor.g;
+  pixelArray[pos + 2] = fillColor.b;
+  pixelArray[pos + 3] = fillColor.a;
+};
+
+const checkColorisNotEqual = (pos: number, startColor: RGBA, pixelArray: Uint8ClampedArray) => {
+  return (
+    pixelArray[pos] !== startColor.r ||
+    pixelArray[pos + 1] !== startColor.g ||
+    pixelArray[pos + 2] !== startColor.b ||
+    pixelArray[pos + 3] !== startColor.a
+  );
+};
+
+export const useDrawingOperation = (
+  canvasRef: RefObject<HTMLCanvasElement>,
+  state: ReturnType<typeof useDrawingState>,
+) => {
+  const { currentColor, brushSize, inkRemaining, setInkRemaining } = state;
+
+  const getCurrentStyle = useCallback(
+    (): StrokeStyle => ({
+      color: currentColor,
+      width: brushSize,
+    }),
+    [currentColor, brushSize],
+  );
+
+  const drawStroke = useCallback((drawingData: DrawingData) => {
+    const { ctx } = getCanvasContext(canvasRef);
+    const { points, style } = drawingData;
+
+    if (points.length === 0) return;
+
+    ctx.strokeStyle = style.color;
+    ctx.fillStyle = style.color;
+    ctx.lineWidth = style.width;
+    ctx.beginPath();
+
+    if (points.length === 1) {
+      const point = points[0];
+      ctx.arc(point.x, point.y, style.width / 2, 0, Math.PI * 2);
+      ctx.fill();
+    } else {
+      ctx.moveTo(points[0].x, points[0].y);
+      points.slice(1).forEach((point) => ctx.lineTo(point.x, point.y));
+      ctx.stroke();
+    }
+  }, []);
+
+  const redrawCanvas = useCallback(() => {
+    if (!state.crdtRef.current) return;
+
+    const { canvas, ctx } = getCanvasContext(canvasRef);
+    ctx.clearRect(0, 0, canvas.width, canvas.height);
+
+    state.crdtRef.current.strokes
+      .filter((stroke) => stroke.stroke !== null)
+      .forEach(({ stroke }) => drawStroke(stroke));
+  }, [drawStroke]);
+
+  const floodFill = useCallback(
+    (startX: number, startY: number) => {
+      const { canvas, ctx } = getCanvasContext(canvasRef);
+      const imageData = ctx.getImageData(0, 0, canvas.width, canvas.height);
+      const pixelArray = imageData.data;
+      const fillColor = hexToRGBA(currentColor);
+
+      const startPos = (startY * canvas.width + startX) * 4;
+      const startColor = {
+        r: pixelArray[startPos],
+        g: pixelArray[startPos + 1],
+        b: pixelArray[startPos + 2],
+        a: pixelArray[startPos + 3],
+      };
+
+      const pixelsToCheck = [[startX, startY]];
+      let pixelCount = 0;
+      const filledPoints: Point[] = [];
+
+      while (pixelsToCheck.length > 0 && pixelCount < inkRemaining) {
+        const [x, y] = pixelsToCheck.shift()!;
+        const pos = (y * canvas.width + x) * 4;
+
+        if (
+          x < 0 ||
+          x >= canvas.width ||
+          y < 0 ||
+          y >= canvas.height ||
+          checkColorisNotEqual(pos, startColor, pixelArray)
+        )
+          continue;
+
+        fillTargetColor(pos, fillColor, pixelArray);
+        filledPoints.push({ x, y });
+        pixelsToCheck.push([x + 1, y], [x - 1, y], [x, y + 1], [x, y - 1]);
+        pixelCount++;
+      }
+
+      ctx.putImageData(imageData, 0, 0);
+      setInkRemaining((prev: number) => Math.max(0, prev - pixelCount));
+
+      return {
+        points: filledPoints,
+        style: getCurrentStyle(),
+      };
+    },
+    [currentColor, inkRemaining, getCurrentStyle, setInkRemaining],
+  );
+
+  return {
+    getCurrentStyle,
+    drawStroke,
+    redrawCanvas,
+    floodFill,
+  };
+};
